05-Digital-Forensics-and-Incident-Response/FileDNA/requirements.txt
pefile
yara-python
ssdeep
pyyaml
rich
